requirements.txt
psycopg2
pandas
tabulate
python-dotenv